refactor: init refactoring of inference & training into seperate distinguishable packages
