+++ backend/suggestion_engine/inference/predict.py
@@ -1,7 +1,7 @@
 import torch
-from suggestion_engine.outcomes.category_suggestion import CategorySuggestion
-from suggestion_engine.outcomes.uncategorized_suggestion import UncategorizedSuggestion
-from suggestion_engine import data
+from suggestion_engine.inference.outcomes.category_suggestion import CategorySuggestion
+from suggestion_engine.inference.outcomes.uncategorized_suggestion import UncategorizedSuggestion
+from suggestion_engine.training.preprocess import data
 import joblib
 import pandas as pd
